WIP - Persist-Settings - Continue Add New ApplicationSettings.
NOTE:  This checkin contains a LOT of commented out code and methods
that will be eventually be removed.

ADD - Add <MyApplicationSettings> argument to the Form1 Ctor.
ADD - Add code to copy the pushSettings obj to the appSettings obj
during the FormClosed event.
ADD - Add property names to the Settings collection.
ADD - Add properties to the MyApplicationsSettings class.  These
properties are equivalent to the properties in the PushSettings class.
DEL - Eliminate a lot of unnecessary code in the Form1( ) Ctor.
DEL - Delete all of the code related to using the old PushSettings and
PushSettingsDefault files in Form1_Load( ).
DEL - Delete all code related to using the old PushSettings files
related to the configuration UI (Form2)
MOD - Add documentation to Itenso.Configuration.dll and
Itenso.Configuration.WindowsForms.dll. No functional change.