WIP - Persist-Settings - BUG Fix
ADD - Add new FormClosing Event handler for Cancel button and 'X'
control on form2.
ADD - Add code to copy the incoming appSettings before they are changed.
This allows the values to be restored if the user selects Cancel button.
MOD - Fix an undiscovered bug with source and target validation and the
error icon.