WIP - Persist-Settings - BUG Fix
Fix issue related to canceling first-run application configuration.  The
user should be able to cancel configuration.  When this happens the
application should shut down, preventing the user from getting to the
main application window.
MOD - Refactor the code that tests to see if the appSettings have ever
been set.
MOD - Change the configuration UI so that the Cancel and 'X' controls
are visible in a first-run scenario.
ADD - Add code to display a message box confirming or canceling the
first-run application setup.
ADD - Add code to form1 to handle the cancel first-run scenario.
ADD - Add code to mail( ) to test for cancel first-run scenario.