Refactor MainForm() Ctor, InitControles(), and UpdateControls()
ADD - Add more tests to IsAppSettingsEmptyOrNull()
ADD - Add code to set a default appSettings value when the current value
is null. Multiple locations.
MOD - Many refactoring changes to MainForm() Ctor, InitControls() and
UpdateControls().
ADD - Add new properties MinShowDetailSize and MinShowDetailSize. The
control the size of the MainForm when it is displaying or hiding the
details.
MOD - Modified bool properties in MyApplicationSettings to bool?.
MISC - Rename a RadioButton that didn't get fixed previously. Rename
form1 to mainForm in Program.cs.