Add Source/Target, File Extension Filter and Dupe File Action to MainForm.
ADD - Add new controls: lblSourcePath, lblTargetPath,
lblFileExtensionFilterString, lblDupeFileActionTest.  The controls will
display the config data on the MainForm.
MOD - Modify the splitter panel layout so that the source and target
path labels are INSIDE the panel.  This makes it easier to handle
resizing the MainForm and keeping the paths in alignment with the
splitter bar.
ADD - Add new hotKey and DEBUG event()
ADD - ToolTip to MainForm. This enables displaying source and target
paths, file filter and Dupe Action mode when hovering over the Show/Hide
label in the MainForm.
ADD - Add code to implement the tooltip function.
MOD - Modify the Regular Expression that splits a file name AAA002 (3)
into a prefix (AAA002) and a interger (3).  The change allowed for 0-*
spaces in between the prefix and the '('.