WIP - Return status counts from Rename, Overwrite, and Skip methods.
NOTE: The returned data is TEST DATA.  It is NOT real data yet.

ADD - Add code to return a Tuple<int,int> object containing the status
counts.  This change was applied to Rename, Overwrite, and Skip.
ADD - Add code to use the returned Tuple to display a simple status
message.
DEL - Remove old references to lbStatus in Rename, Overwrite, and Skip
methods.