WIP Progress Bar - Implement Framework and CopyRename
MOD - Show the progress bar and progress label in the
ProgressChangedEventHandler.
MOD - Update the Progress bar in the ProgressChangedEventHandler.
ADD - Add the WorkerCompletedEventHandler to the MainForm.  Completely
re-write the event handler.
ADD - Add code to hide the status labels, progress label, and progress
bar in InitControls().
MOD - Modify the background event handler called for
RunWorkerCompletedEventHadler.
MOD - Modify the DoWorkEventHandler for CopyOverWrite.  This is
temporary.
MOD - Modify the DoWorkEventHandler for CopyRename. Add a collection of
arguments to the event handler. Change the name of the event handler to
CopyFileRename.RenameDuplicates_BackGround. This will be renamed later
as part of a cleanup effort.
ADD - Create a new version of RenameDuplicates called
RenameDuplicates_BackGround( ).  This version includes additional code
to extract the arguments passed in the DoWorkEventArgs.  This version
also assigns the processing results to the DoWorkEventsArgs.Result
property.  The Results property contains the processing counts that are
displayed in the UI.  This version also sets the ReportProgress value
that is used to update the progress bar.

NOTE: A great deal of exploratory code has been deleted in this check
in.  --- This check in deals with the CopyRename behavior only.  The
CopyOverWrite and CopySkip, and Cancel behaviors will be next. --- Many
of the variables and methods will be renamed in future check ins, E.G.
CopyFileRename.RenameDuplicates_BackGround( ).  Old code will be purged.