WIP Progress Bar - Implement SkipDuplicates (Manual)
MOD - Change the Copy Overwrite event handler from
bgDoWorkEventHandlerSkipDuplicates( ) to
CopyFileOverwrite.SkipDuplicates_BackGround( ).
ADD - Add new method SkipDuplicates_BackGround( ) based on
SkipDuplicates( ). This version includes additional code to extract the
arguments passed in the DoWorkEventArgs.  This version also assigns the
processing results to the DoWorkEventsArgs.Result property.  The Results
property contains the processing counts that are displayed in the UI.
This version also sets the ReportProgress value that is used to update
the progress bar.