BUG - Progress Bar does not go to 100%
MOD - Move the progress bar to the toolbar.
MOD - Mod the progress label to the toolbar.
DEL -  Code that hides the progress bar when Backgroundworker completes.
This was preventing the UI from being updated.  The values WAS actually
getting to 100%.
ADD - Add code to hide progress bar and label when toolbar - refresh is
selected.
ADD - Add code to hide progress bar and label when the PUSH button is
selected.  This resets the the UI for the next batch of processing.