BUG - When there are ZERO files in the source folder an incorrect message is displayed "0 Files Overwritten".
ADD - Add new test for ZERO files in source folder.
ADD - Add BackGround Event handler that just returns the
commandResult.Fail value to the bgRunWorkerCompletedEventHandler.  This
causes the UpdateStatus( ) method to be called.
ADD - Add new case for Helper.commandResult.Fail.  This creates the
correct message to be displayed in the UI: "No Files to Copy".