Resize Main UI
ISSUE - Enlarge the height of the Main UI when in details are visible.
Should be able to see 4x lines in ListViews.

ISSUE - Enlarge the width of the Main UI when the details are hidden.
Need enough room to see "xxx Files Overwritten" with enough margin to
right of label.

MOD - Modify the Min Hide Detail size width.
MOD - Modify the Min Show Detail Size height.