ISSUE - Sort-Source/Target-ListView by FileName, Type, Date
NOTE: The ListViewColumnSorter class was written by Sven So.  It was
published on Code Project. I did not create this class.  I implemented
it in the application.
ADD - Add new file, ListViewColumnSort.cs.  This file contains the class
ListViewColumnSorter.
ADD - Add event handler ListView_ColumnClick.  This handler supports
source and target ListViews.
ADD - Add code to MainForm Ctor to initialize the source and target
ListViewItemSorter property with a ListViewColumnSorter type.