ISSUE - Expand the FileName column to fit the longest name on-demand.
ADD - Add TWO controls to the Source and target panel.
ADD - Add event handlers to support the new controls.  These handlers
set the FileName column width to fit the widest value in the column.
ADD - Add new Fit.gif to the project and to the project/resources
folder.  This gif is used in the TWO new controls.