Create missing subfolders in target folder + Copy files into subfolders
ADD - Add code to CopyOverwrite( ) to find all unique subfolders in the
source file list.
ADD - Add code to CopyOverwrite( ) to create any paths in target folder
that doesn't exist.
ADD - CreateTargetPaths( ). This method checks for each subfolder in the
target and creates them. This method is recursive.
MOD - Modify CopyOverwrite( ) to copy files in to the proper subfolder.