Add Ignore Files Feature. (Restore)
COMMENT --- This check in contains WORKING CODE.  It however, has not
been cleaned or refactored.  The next check in will be BASIC code
cleanup with NO functional change.

COMMENT: This code was originally checked in on 123015 but was
misplaced. This is the missing code.
ADD - Add new file: Config\PushIgnore.txt. This file contains a list of
files that should be ignored when processing.  The files is based on
.gitIgnore. It is used to create a RegEx pattern that performs
filtering.  This file is not used as a substitute to the Configure >
File Extensions.
MOD - Modify the GetFiles( ) support ignorePattern.
ADD - Add new method BuildIgnorePattern( ). This method reads the
PushIgnore.txt file and generates a RegEx pattern.  This method is
called during the construction of MainForm during the InitControls( )
stage.