ISSUE - If Copy - FileOverwrite fails, do not Delete source
COMMENT -- When a copy fails the application should generate a
descriptive message and exit the action.  When the copy is successful
but the cleanup/Delete fails the application should remove the new
copied files. When a document exists in the source and target the
application must insure that the original, pre-copy, file is restored.

MOD - Modify bgRunWorkerCompletedEventHandler( ) to test for a returned
exception and generate a message for the user.
ADD - Add new function Helper.CopyDelete( ). This function encapsulates
the Copy and Delete functions and wraps them in Try/Catch blocks.
MOD - Modify CopyOverwrite_BackGround( ) to call Helper.CopyDelete( ).