ISSUE - If Copy - FileSkip or Copy - Rename fails, do not Delete source
COMMENT -- When a copy fails the application should generate a
descriptive message and exit the action.  When the copy is successful
but the cleanup/Delete fails the application should remove the new
copied files. When a document exists in the source and target the
application must insure that the original, pre-copy, file is restored.

MOD - Add default argument to CopyDelete( ) such that the default action
with be to overwrite the target file if it exists.
MOD - Modify SkipDuplicates_BackGround( )  to call CopyDelete( ). The
default argument is used.
MOD - Modify RenameDuplicates_BackGround( ) to call CopyDelete( ).
Override the default argument to false (Do not overwrite exiting file).