ISSUE - User.Conf - invalid XML causes app to fail to start.
COMMENT:  User.Conf - The appSettings.Load( ) throws an exception
because of invalid XML. Modify the code to catch the exception and
delete the existing user.config file. The application will then
regenerate the file the next time it exits.

MOD - Modify Main( ) to trap appSettings.Load(), delete invalid
user.config file.
MOD - Modify the ValidateAppSettings( ). The AppSettingsEmptyOrNull( )
was returning the wrong value.
MOD - Modify BuildIgnorePattern( ) to fetch the Push.exe path when the
appSettings.ExePath is null.
MOD - Remove the HideDupeMessage and ShowDetails test. This values are
set properly later in the application.