ISSUE - PushConfig + Temp File Name - Remove the  Remove Max/Min/Resize controls + Rename Temp to PushTemp
COMMENT:  PushConfig Dialog - The user should not be allowed to use the
Min/Max/Resize controls on this dialog.

COMMENT:  The TEMP filename, used for 'undoing' copies after exceptions
is not specific to the application.  Rename the file PushTemp and add
code to clean up the file after a successful copy.

MOD - Rename the temp file name from 'TEMP' to 'PushTemp'
MOD - Add Finally block to clean up the PushTemp file when it exists.
ADD -  Add code to 'lock' the ConfigForm and hide the Minimize/Maximize
controls in the Form.