ISSUE - [Ctrl-D,0] trys to delete all files in the source folder. This deletes the ALL the files. Redesign Debug Hot-Keys
COMMENT: This issue is not a problem that a user should encounter. It
may be encountered if someone is trying to debug a problem. This change
re-designs the way that the Debug Hot-Key functions. There are certain
data structures that must be present for the Hot-Key to function. A
special DEGUG.zip file helps set this structure up.  The DEGUG.zip file
'may' be added to the repo along with some additional notes at a later
date.

MOD - DEBUG_Red( ) // HotKey 0. Change the location where the debug data
and source folders are expected to be.
ADD - DEBUG_Red( ) // HotKey 0. Add code to notify user when one of the
dependencies is not available.
MOD - DEBUG_Red( ) // HotKey 0. Modify call to initialize/clear all data
to the SRC & DEST folders.
MOD - DEBUG_Red( ) // HotKey 0. Change the location where the debug data
and source folders are expected
to be.
ADD - DEBUG_Red( ) // HotKey 0. Add Code to validate the folders that
the Hot-Key depends on.
MOD - DEBUG_Red( ) // HotKey 0. Modify calls to load test data into the
SRC and DEST folders.

MOD - IsValidDebugDependencies( ). Rewrite the way the folder
dependencies are validated.

MOD - DEBUG_InitFolders( ). Modify code to use the new Dictionary object
for clearing the SRC and DEST folders.